First begin with documentation
This is a first draft of a first begin of documentation.  Currently it's
rather a rough collection of notes and hypothesis of what the purpose is
of the various scripts in this package.  I hope this start may encourage
developers with a better understanding of this package to expand the
documentation.